requirements.txt
pandas
sklearn
nltk
django
djangorestframework
requests
gunicorn
whitenoise
dj-database-url